Add files via upload
Here are the notebooks I have been working on. I need to clean these up more but I wanted to share enough for us to get started. The hope is that we can use the manual coded data to create an NLP model. Unfortunately, the main dataset was too large to upload here, so if you want to program these you will need to download that from Kaggle. Sorry for the inconvenience.